Expose declassified configuration
