+++ package.json
@@ -26,14 +26,14 @@
         "yarn": "^1.5.1"
     },
     "peerDependencies": {
-        "@globality/nodule-config": ">= 2.9.0 < 3",
+        "@globality/nodule-config": ">= 2.11.0 < 3",
         "@globality/nodule-openapi": ">= 0.9.3 < 1",
-        "@globality/nodule-logging": ">= 1.2.0 < 2"
+        "@globality/nodule-logging": ">= 1.4.0 < 2"
     },
     "devDependencies": {
-        "@globality/nodule-config": "~2.9.0",
+        "@globality/nodule-config": "~2.11.0",
         "@globality/nodule-openapi": "~0.9.3",
-        "@globality/nodule-logging": "~1.2.2",
+        "@globality/nodule-logging": "~1.4.0",
         "babel-cli": "^6.26.0",
         "babel-eslint": "^8.2.2",
         "babel-plugin-transform-async-to-generator": "^6.24.1",
@@ -49,7 +49,7 @@
         "supertest": "^3.0.0"
     },
     "jest": {
-        "testRegex": "/__tests__/.*(test)\\.js",
+        "testRegex": "/__tests__/.*test\\.js$",
         "modulePaths": [
             "<rootDir>/src"
         ]
